wechat/requirements.txt
aiohttp==3.0.6
asn1crypto==0.24.0
astroid==1.5.3
async-timeout==2.0.0
attrs==17.3.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bleach==2.1.2
boost==0.1
bootstrap-admin==0.3.7.1
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
cmake==0.9.0
colorama==0.3.9
Crawl==0.5.5
cryptography==2.1.4
cycler==0.10.0
DataFile==0.1.6
decorator==4.1.2
Django==1.11.7
django-haystack==2.6.1
django-proxy==1.1.0
echarts==0.0.0
echarts-python==0.1.3
entrypoints==0.2.3
Fabric==1.14.0
fire==0.1.2
freeze==1.0.10
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
html5lib==1.0.1
idna==2.5
idna-ssl==1.0.0
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
isort==4.2.15
itchat==1.3.10
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter-echarts-pypkg==0.1.0
Keras==2.1.4
lazy-object-proxy==1.3.1
lml==0.0.2
lulu==0.4.8
lxml==4.1.1
Markdown==2.6.8
MarkupSafe==1.0
Mastodon.py==1.2.1
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.8.3
multidict==4.1.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
opencv-python==3.4.0.12
pandas==0.21.0
pandocfilters==1.4.2
paramiko==2.4.0
parso==0.1.1
peppercorn==0.5
pew==1.1.2
pickleshare==0.7.4
Pillow==4.3.0
pipenv==11.0.2
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.3.1
py==1.5.2
py2exe==0.9.2.2
pyasn1==0.4.2
pycparser==2.18
pyecharts==0.4.1
pyecharts-jupyter-installer==0.0.3
pyexcel==0.5.6
pyexcel-io==0.5.4
pyexcel-text==0.2.7.1
Pygments==2.2.0
pylint==1.7.1
pymongo==3.5.1
PyNaCl==1.2.1
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
selenium==3.8.0
simplegeneric==0.8.1
six==1.11.0
splinter==0.7.7
SQLAlchemy==1.2.0
tabulate==0.8.2
testpath==0.3.1
texttable==1.1.1
tornado==4.5.2
torrent-cli==0.0.3
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Whoosh==2.7.4
widgetsnbextension==3.1.0
WMI==1.4.9
wordcloud==1.4.1
wrapt==1.10.11
xlrd==1.1.0
yarl==1.1.1
